dell::openmanage::redhat: resolve dependency cycle
There was a dependency cycle when installing redhat on physical servers.
Removing that dependency breaks that cycle.

I'm not exactly sure why this dependency was there in the first place. It seems
weird that a repository definition depends on a package. The openmanage recipes
seem to apply just fine without this require.