dell::openmanage: kill probable dependency cycle
We have a general puppet rule that packages depend on yum cache cleaning.
So we can't clean yum cache after installing a package that depends on that
yum cache cleaning. Anyway, it's fine if that plugin is disabled for the next
puppet run or user action. We don't need yum to clean its cache right now so
we can safely get rid of that notification.